remove intl
Summary:
`intl` is deprecated in the latest arrayfire version. Mostly replace with `int` which is suitable for 99% of cases. In the other 1% we can use `int64_t` or ArrayFire's `dim_t` instead.

For class data members which used to be `intl`, we replace them with `int64_t` to avoid breaking serialization.

Note that for serialization, we do assume that `sizeof(int) == 4` across all relevant platforms. Otherwise it would be very inconvenient. This is a pretty safe and widespread assumption nowadays.

Reviewed By: vineelpratap

Differential Revision: D13716651

fbshipit-source-id: 4a66eb69c12bbbbdf5895ae8bb1bc35c46fdc02b