Train.cpp for semisupervised exps
Summary:
Copied `wav2letter/Train.cpp` and made a clean version under `wav2letter/src/experimental/semisupervised` for seq2seq semisupervised experiments in the future.

Main differences to the original `Train.cpp` are:
1. Move all lambda expressions except `train` to separate files under `wav2letter/src/experimental/semisupervised/runtime` as functions for better readability. Decide to keep `train()` as lambda in the main file as it might be easier to set up schedules to train different components in the future.
2. Use one optimizer for both network & criterion, since we've been using the same lr for both.
3. Set up a `LogHelper` to take care of all logging to files and model saving.
4. Merge the two checkpoints at the end of each epoch and at a fixed number of iterations into one checkpoint.

Some of the changes can be applied to the original `Train.cpp` as well. Happy to cover some of that if people have a consensus.

Next steps:
1. More flexible logging
2. Extra dataset loading
3. Extra module for LM

Reviewed By: awni

Differential Revision: D14917383

fbshipit-source-id: cc53d86141f96935270f8059418a1e63ddae56c7