Update Test.cpp to make sample numbers in output consistent with Decode
Test.cpp produces predictions of characters without the LM. 
Without the proposed change, the sample numbers in the output just progress from 1 to n (where n is the number of files in the test set). This is inconsistent with the Decode ouput which lists sample number.
The proposed change make the "sample" output consistent with the Decode output. This is useful if you want to see where the language model is modifying the predictions of the Acoustic model