Add Reducer framework for distributed optimization
Summary:
Adds scaffolding for a framework for writing custom gradient synchronization/reduction algorithms in flashlight and adds it to existing autograd gradient sync. Implements the existing single-stream, synchronous `allReduce` as a `PassthroughReducer` which is selected by default

The goal here is to allow experimentation with different parallelism and parameter update rules and to expose a generic API that can be used in the training flow (for instance, my best-performing distributed training configuration requires calling a `synchronize()`-like function after `Variable::backward` returns to ensure the AF and NCCL CUDA streams are in sync.

The `Reduce` API may change as we add more complex reducers don't the line; this is what I anticipate it will look like after implementing coalescing.

Not planning on landing this until I've implemented other reducers (e.g. coalescing-based reducers) and `allReduceSet` on the distributed API side.

Reviewed By: vineelpratap

Differential Revision: D14739285

fbshipit-source-id: 3f3c2cf7ceac7511e58509894a5b4cf34ef14d22