Fix segfault with no --enable_distributed
Summary:
In a single-node run when `--enable_distributed` was not specified, `initWithMPI()` was not called, causing the `cudaEvent_t` of the `CoalescingReducer` to be `nullptr`. This crashed the program when it called `syncDistributed()`.

The most appropriate solution is that if `--enable_distributed` is not specified, then:
- `fl::distributedInit()` should never be called (already the case)
- `fl::Reducer` should never even be created
- `fl::distributeModuleGrads()` should never be called
- `fl::Reducer::finalize()` should never be called

This diff implements these items.

Reviewed By: jacobkahn

Differential Revision: D15506108

fbshipit-source-id: b463f9c88b832aa9ff685bce5e6ee4f0cad7b04e