create runtime/Helpers.h, remove randomSubset()
Summary:
I felt that we needed a place to put reusable utility functions for Train.cpp and other binaries, where those functions are not generic enough to be in `common/Utils.h` or similar. So I came up with `runtime/Helpers.h`.

The `randomSubset()` function in `common/Utils.h` is one such example and the one I started out with. Subsequent diffs in stack show examples of putting more stuff in `runtime/Helpers.h`.

Reviewed By: jacobkahn

Differential Revision: D15891460

fbshipit-source-id: 0bcfe93409ebb3c0a58449cf5ab528ee5aa5984f