Add nDecodedFramesInBuffer() into the base decoder class
Summary:
Add one function which is implemented in each sub-decoder classes but not appeared in the base class.
`nDecodedFramesInBuffer()` is a function showing the number of frames that are decoded and stored in buffer. It is really important to have it in the base class to work with `prune()` together.

Reviewed By: vineelpratap

Differential Revision: D16881075

fbshipit-source-id: d0144072918fc1ca825156356f39b69e32c8428b