Add memcache support in front of Everstore
Summary:
Adds support for Memcache to cache Everstore handles and ranges in front of `W2lEverstoreReader`, and adds `--use_memcache` for internal runs, which queries the `datatype::MemcacheDataType` accordingly.

Cache lookups are only supports cache lookups for single handles - doesn't support multi-handle lookup. There's apparently no benefit to batch lookups in the cache, so we'll stick with this for now (this is also what `W2lEverstoreReader` expects, since each thread reads a handle.

We'll do some diagnostics/performance analysis before landing and after D17865315 is committed.

### Cache Keys

If the key has an offset and a length:
```
wav2letter:<everstore handle>:<audio offset>:<audio length>:c1
```
if the key has no offset/length:
```
wav2letter:audio:<everstore handle>:c1
```

Since we'll never read overlapping regions of audio from Everstore or the cache (everything is disjoint), it's most efficient to cache the range read from Everstore using a key that's range-aware, since that's how we'll interact with that data in the future.

Reviewed By: edenzik

Differential Revision: D17865397

fbshipit-source-id: e4d4891e8b9f95846b0768116848ef95cace4328