Fixed bug causing "Unknown speaker!"
Judging by assertion (line 115) and speaker access (line 117) statements, I think on line 111 the intention was to create a dictionary from the list of speakers {speaker.id: speaker for speaker in speakers}, but instead a set of tuples {(speaker.id, speaker) for speaker in speakers} was created, causing assert speaker in speakers to fail. I fixed it accordingly.