adding the mtl loss
Summary:
Adding the MTL loss to the joint_training recipe.

The MTL loss force the ASR model to predict a given label for each sequence on top of the usual decoding. For example, in a multilingual scenario, the MTL Loss would be used to predict the target language of the sample.

If the MTL loss is activated, then each file ID in the train.lst should have the following format: fileID = "{baseID}#{MTL Loss label}". For example file1#en. The list of the possible labels should be provided in a separate file given with the command line. See joint_training_voxpopuli/cpc/test/test_mapping.txt for the file format.

To activate the MTL Loss add --mtllossmapping $LABELFILE to your command line. You can also change the weight of the loss with the argument --mtllossweight.

Reviewed By: syhw

Differential Revision: D31651548

fbshipit-source-id: c2c123a5a9b65ac932814a703011ff1a616590dc